remove fileDownlaod from RequestPerformable & AsyncRequestPerformable + unit tests
