+++ Sources/EZNetworking/Util/Performers/RequestPerformer.swift
@@ -4,7 +4,6 @@ import UIKit
 public protocol RequestPerformable {
     func performTask<T: Decodable>(request: URLRequest, decodeTo decodableObject: T.Type, completion: @escaping((Result<T, NetworkingError>)) -> Void) -> URLSessionDataTask
     func performTask(request: URLRequest, completion: @escaping((VoidResult<NetworkingError>) -> Void)) -> URLSessionDataTask
-    func downloadFileTask(url: URL, completion: @escaping((Result<URL, NetworkingError>) -> Void)) -> URLSessionDownloadTask
     func downloadImageTask(url: URL, completion: @escaping((Result<UIImage, NetworkingError>) -> Void)) -> URLSessionDataTask
 }
 
@@ -51,19 +50,6 @@ public struct RequestPerformer: RequestPerformable {
             } catch {
                 completion(.failure(NetworkingError.unknown))
                 return
-            }
-        }
-    }
-    
-    public func downloadFileTask(url: URL, completion: @escaping((Result<URL, NetworkingError>) -> Void)) -> URLSessionDownloadTask {
-        return urlSession.downloadTask(with: url) { localURL, response, error in
-            do {
-                let localURL = try urlResponseValidator.validateDownloadTask(url: localURL, urlResponse: response, error: error)
-                completion(.success(localURL))
-            } catch let networkError as NetworkingError {
-                completion(.failure(networkError))
-            } catch {
-                completion(.failure(.unknown))
             }
         }
     }
