+++ Tests/EZNetworkingTests/Util/Performers/AsyncRequestPerformableTests.swift
@@ -192,84 +192,6 @@ final class AsyncRequestPerformableTests: XCTestCase {
         }
     }
     
-    // MARK: Unit tests for downloadFile
-
-    func testDownloadFileSuccess() async throws {
-        let testURL = URL(string: "https://example.com/example.pdf")!
-        let urlSession = MockURLSession(
-            url: testURL,
-            urlResponse: buildResponse(statusCode: 200),
-            error: nil
-        )
-        let validator = MockURLResponseValidator(throwError: nil)
-        let decoder = RequestDecoder()
-        let sut = AsyncRequestPerformer(urlSession: urlSession, urlResponseValidator: validator, requestDecoder: decoder)
-        
-        do {
-            let localURL = try await sut.downloadFile(with: testURL)
-            XCTAssertEqual(localURL.absoluteString, "file:///tmp/test.pdf")
-        } catch {
-            XCTFail()
-        }
-    }
-    
-    func testDownloadFileFailsWhenValidatorThrowsAnyError() async throws {
-        let testURL = URL(string: "https://example.com/example.pdf")!
-        let urlSession = MockURLSession(
-            url: testURL,
-            urlResponse: buildResponse(statusCode: 200),
-            error: nil
-        )
-        let validator = MockURLResponseValidator(throwError: NetworkingError.httpError(.forbidden))
-        let decoder = RequestDecoder()
-        let sut = AsyncRequestPerformer(urlSession: urlSession, urlResponseValidator: validator, requestDecoder: decoder)
-        
-        do {
-            _ = try await sut.downloadFile(with: testURL)
-            XCTFail("unexpected error")
-        } catch let error as NetworkingError{
-            XCTAssertEqual(error, NetworkingError.httpError(.forbidden))
-        }
-    }
-    
-    func testDownloadFileFailsWhenStatusCodeIsNot200() async throws {
-        let testURL = URL(string: "https://example.com/example.pdf")!
-        let urlSession = MockURLSession(
-            url: testURL,
-            urlResponse: buildResponse(statusCode: 400),
-            error: nil
-        )
-        let validator = URLResponseValidatorImpl()
-        let decoder = RequestDecoder()
-        let sut = AsyncRequestPerformer(urlSession: urlSession, urlResponseValidator: validator, requestDecoder: decoder)
-        
-        do {
-            _ = try await sut.downloadFile(with: testURL)
-            XCTFail("unexpected error")
-        } catch let error as NetworkingError{
-            XCTAssertEqual(error, NetworkingError.httpError(.badRequest))
-        }
-    }
-    
-    func testDownloadFileFailsWhenErrorIsNotNil() async throws {
-        let testURL = URL(string: "https://example.com/example.pdf")!
-        let urlSession = MockURLSession(
-            url: testURL,
-            urlResponse: buildResponse(statusCode: 200),
-            error: NetworkingError.unknown
-        )
-        let validator = URLResponseValidatorImpl()
-        let decoder = RequestDecoder()
-        let sut = AsyncRequestPerformer(urlSession: urlSession, urlResponseValidator: validator, requestDecoder: decoder)
-        
-        do {
-            _ = try await sut.downloadFile(with: testURL)
-            XCTFail("unexpected error")
-        } catch let error as NetworkingError{
-            XCTAssertEqual(error, NetworkingError.unknown)
-        }
-    }
-    
     
     // MARK: Unit tests for downloadImage
